Filering Bridge

+++ executor/coordinator/test/integration_preparer_test.sh
@@ -7,8 +7,10 @@ set -e
 MYDIR=`dirname $0`
 cd $MYDIR/..
 
+MANTIKFILE="IyBUaGlzIGlzIGEgZHVtbXkgZmlsZSBmb3IgdGVzdGluZyB0aGUgcHJlcGFyZXItQ29udGFpbmVyCm5hbWU6IEhlbGxvIFdvcmxkCnZlcnNpb246IDAuMQo="
+
 # Base64 made at command line with base64 utility.
-./target/payload_preparer -url file://test/hello1.zip -dir test/unpack_test -pdir sub -mantikfile "IyBUaGlzIGlzIGEgZHVtbXkgZmlsZSBmb3IgdGVzdGluZyB0aGUgcHJlcGFyZXItQ29udGFpbmVyCm5hbWU6IEhlbGxvIFdvcmxkCnZlcnNpb246IDAuMQo="
+./target/payload_preparer -url file://test/hello1.zip -dir test/unpack_test -pdir sub -mantikfile $MANTIKFILE
 
 if [ ! -f test/unpack_test/sub/hello1/HelloWorld.txt ]; then
     echo "Unpacked bundle does'nt exist"
@@ -25,5 +27,14 @@ if ! diff test/unpack_test/Mantikfile test/hello1_mantikfile.yaml; then
     exit 1
 fi
 
-echo "All fine, deleting temporary directory"
+echo "Deleting temporary directory"
+rm -r test/unpack_test
+
+# 2nd test with just Mantikfile and no payload
+./target/payload_preparer -dir test/unpack_test -mantikfile $MANTIKFILE
+if [ ! -f test/unpack_test/Mantikfile ]; then
+    echo "Mantikfile doesn't exist"
+    exit 1
+fi
+echo "Deleting temporary directory"
 rm -r test/unpack_test
